requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
click==8.3.2
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.135.3
greenlet==3.4.0
groq==1.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
packaging==26.0
pydantic==2.13.0
pydantic_core==2.46.0
python-dotenv==1.2.2
python-multipart==0.0.26
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
psycopg2-binary==2.9.11
